+++ R/rstudio.R
@@ -5,11 +5,12 @@
 #' @param instance An instance with RStudio installed via \link{gce_vm_template}
 #' @param username The user to create
 #' @param password The user password
+#' @param admin Default TRUE - Will the user be able to install packages and other sudo tasks?
 #' @param container The rstudio container to add the user to
 #' 
 #' @return The instance
 #' @export
-gce_rstudio_adduser <- function(instance, username, password, 
+gce_rstudio_adduser <- function(instance, username, password, admin = TRUE,
                                 container = "rstudio"){
   
   ssh_au <- paste0("adduser ",
@@ -20,6 +21,13 @@ gce_rstudio_adduser <- function(instance, username, password,
                      cmd = "exec",
                      args = c(container, ssh_au),
                      docker_opts = "")
+  
+  if(admin){
+    docker_cmd(instance, cmd = "exec",
+               args = c(container, 
+                        paste0("sudo adduser ",username," staff"))
+    )
+  }
   
   docker_cmd(instance,
                      "exec",