+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - Update website
 - Add R-Datalab Dockerfile example
+- Let Rstudio users be added with staff rights so they can install packages etc. 
 
 # googleComputeEngineR 0.1.0